# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Manuskript authoring and Story Oracle API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /video/projects/{projectId}/assets/{assetId}/reference-analyze:
    post:
      operationId: analyzeVideoReference
      tags: [content-creators]
      summary: Queue pacing analysis of a REFERENCE asset (viral reference import)
      parameters:
        - { $ref: "#/components/parameters/VideoProjectId" }
        - { $ref: "#/components/parameters/VideoAssetId" }
      responses:
        "201":
          description: Queued analysis job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VideoJob" }
        "403":
          description: Not allowed (not the SELECTS role / Captain)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /video/projects/{projectId}/assets/{assetId}/reference:
    get:
      operationId: getVideoReference
      tags: [content-creators]
      summary: Read the extracted pacing structure of a reference asset
      parameters:
        - { $ref: "#/components/parameters/VideoProjectId" }
        - { $ref: "#/components/parameters/VideoAssetId" }
      responses:
        "200":
          description: Reference pacing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VideoReference" }
        "404":
          description: No analysis yet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Not allowed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /video/projects/{projectId}/grants:
    get:
      operationId: listVideoGrants
      tags: [content-creators]
      summary: List temporary download grants (Captain)
      parameters:
        - { $ref: "#/components/parameters/VideoProjectId" }
      responses:
        "200":
          description: Grants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/VideoGrant" }
        "403":
          description: Captain only
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    post:
      operationId: createVideoGrant
      tags: [content-creators]
      summary: Grant a member a temporary download (Captain)
      parameters:
        - { $ref: "#/components/parameters/VideoProjectId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoGrantInput" }
      responses:
        "201":
          description: Created grant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VideoGrant" }
        "400":
          description: Invalid grant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Captain only
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /video/projects/{projectId}/grants/{grantId}/revoke:
    post:
      operationId: revokeVideoGrant
      tags: [content-creators]
      summary: Revoke a temporary download grant instantly (Captain)
      parameters:
        - { $ref: "#/components/parameters/VideoProjectId" }
        - { $ref: "#/components/parameters/VideoGrantId" }
      responses:
        "200":
          description: Revoked grant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VideoGrant" }
        "403":
          description: Captain only
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Grant not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /video/notifications/{notificationId}/read:
    post:
      operationId: markVideoNotificationRead
      tags: [content-creators]
      summary: Mark a notification as read
      parameters:
        - { $ref: "#/components/parameters/VideoNotificationId" }
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VideoNotification" }
        "404":
          description: Notification not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
components:
  parameters:
...
    VideoProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
    VideoAssetId:
      name: assetId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
...
    VideoGrantId:
      name: grantId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
    VideoNotificationId:
      name: notificationId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    VideoJob:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        assetId: { type: string }
        type: { type: string }
        status: { type: string }
        attempts: { type: integer }
        error: { type: ["string", "null"] }
        params: { type: ["object", "null"], additionalProperties: true }
        result: { type: ["object", "null"], additionalProperties: true }
        createdAt: { type: string, format: date-time }
        startedAt: { type: ["string", "null"], format: date-time }
        finishedAt: { type: ["string", "null"], format: date-time }
      required: [id, projectId, assetId, type, status, attempts, error, params, result, createdAt, startedAt, finishedAt]
...
    VideoReference:
      type: object
      properties:
        id: { type: string }
        assetId: { type: string }
        status: { type: string }
        pacing: { type: ["object", "null"], additionalProperties: true }
        error: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, assetId, status, pacing, error, createdAt, updatedAt]
    VideoGrant:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        fileId: { type: string }
        memberId: { type: string }
        reason: { type: string }
        grantedById: { type: string }
        expiresAt: { type: string, format: date-time }
        revokedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
      required: [id, projectId, fileId, memberId, reason, grantedById, expiresAt, revokedAt, createdAt]
    VideoGrantInput:
      type: object
      properties:
        memberId: { type: string, minLength: 1 }
        fileId: { type: string, minLength: 1 }
        reason: { type: string, maxLength: 500 }
        expiresInHours: { type: integer, minimum: 1, maximum: 168, default: 24 }
      required: [memberId, fileId]
    VideoNotification:
      type: object
      properties:
        id: { type: string }
        recipientId: { type: string }
        category: { type: string }
        title: { type: string }
        body: { type: string }
        deepLink: { type: string }
        resourceId: { type: ["string", "null"] }
        readAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
      required: [id, recipientId, category, title, body, deepLink, resourceId, readAt, createdAt]