# lib/api-zod/src/generated/types/videoGrant.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Manuskript authoring and Story Oracle API
 * OpenAPI spec version: 0.2.0
 */

export interface VideoGrant {
  id: string;
  projectId: string;
  fileId: string;
  memberId: string;
  reason: string;
  grantedById: string;
  expiresAt: Date;
  /** @nullable */
  revokedAt: Date | null;
  createdAt: Date;
}